Reformat project; apply code style

+++ BepInEx.Core/Console/Windows/Kon.cs
@@ -102,7 +102,7 @@ namespace BepInEx.ConsoleUtil
             if (!flag)
                 return;
             var num = bufferInfo.wAttributes;
-            num &= isBackground ? (short) -241 : (short) -16;
+            num &= (short) (isBackground ? -241 : -16);
             num = (short) ((ushort) num | (ushort) color);
             SetConsoleTextAttribute(conOut, num);
         }
